workflows/requirements.txt
-c constraints.txt
--extra-index-url https://download.pytorch.org/whl/cpu
inspire-utils~=3.0.68
idutils
Sickle==0.7.0
plotextractor==1.0.22
invenio-classifier==1.3.13
inspire-schemas==61.6.26
refextract==1.1.6
inspire-json-merger==11.0.44
inspire-classifier==2.0.5
apache-airflow-providers-opensearch
